[text-to-vector-refactoring] Post check changes, added changelog and fixed nomenclature
