+++ solr/modules/language-models/src/java/org/apache/solr/languagemodels/package-info.java
@@ -15,5 +15,5 @@
  * limitations under the License.
  */
 
-/** Contains model store related classes. */
-package org.apache.solr.languagemodels.store;
+/** Contains the whole module for Language Models. */
+package org.apache.solr.languagemodels;
