--- solr/modules/language-models/src/test-files/embeddingModelExamples/dummy-model.json
@@ -1,7 +1,0 @@
-{
-  "class": "org.apache.solr.languagemodels.textvectorisation.model.DummyEmbeddingModel",
-  "name": "dummy-1",
-  "params": {
-    "embedding": [1.0, 2.0, 3.0, 4.0]
-  }
-}
