--- solr/modules/language-models/src/test-files/embeddingModelExamples/exception-throwing-model.json
@@ -1,6 +1,0 @@
-{
-  "class": "org.apache.solr.languagemodels.textvectorisation.model.ExceptionThrowingEmbeddingModel",
-  "name": "exception-throwing-model",
-  "params": {
-  }
-}
