+++ solr/modules/language-models/src/test-files/solr/collection1/conf/solrconfig-language-models.xml
@@ -55,7 +55,7 @@
  </requestHandler>
 
  <updateRequestProcessorChain name="textToVector">
-  <processor class="solr.languagemodels.textvectorisation.update.processor.TextToVectorUpdateProcessorFactory">
+  <processor class="solr.languagemodels.update.processor.factory.TextToVectorUpdateProcessorFactory">
    <str name="inputField">_text_</str>
    <str name="outputField">vector</str>
    <str name="model">dummy-1</str>
@@ -64,7 +64,7 @@
  </updateRequestProcessorChain>
 
  <updateRequestProcessorChain name="failingTextToVector">
-  <processor class="solr.languagemodels.textvectorisation.update.processor.TextToVectorUpdateProcessorFactory">
+  <processor class="solr.languagemodels.update.processor.factory.TextToVectorUpdateProcessorFactory">
    <str name="inputField">_text_</str>
    <str name="outputField">vector</str>
    <str name="model">exception-throwing-model</str>
@@ -73,7 +73,7 @@
  </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="textToVectorStoredInputField">
-    <processor class="solr.languagemodels.textvectorisation.update.processor.TextToVectorUpdateProcessorFactory">
+    <processor class="solr.languagemodels.update.processor.factory.TextToVectorUpdateProcessorFactory">
       <str name="inputField">string_field</str>
       <str name="outputField">vector</str>
       <str name="model">dummy-1</str>
@@ -83,7 +83,7 @@
 
   <updateRequestProcessorChain name="textToVectorForPartialUpdates">
     <processor class="solr.DistributedUpdateProcessorFactory"/>
-    <processor class="solr.languagemodels.textvectorisation.update.processor.TextToVectorUpdateProcessorFactory">
+    <processor class="solr.languagemodels.update.processor.factory.TextToVectorUpdateProcessorFactory">
       <str name="inputField">string_field</str>
       <str name="outputField">vector</str>
       <str name="model">dummy-1</str>
