+++ solr/modules/language-models/src/test-files/textToVectorModelExamples/cohere-model.json
