+++ solr/modules/language-models/src/test-files/textToVectorModelExamples/dummy-model-ambiguous.json
@@ -1,5 +1,5 @@
 {
-  "class": "org.apache.solr.languagemodels.textvectorisation.model.DummyEmbeddingModel",
+  "class": "org.apache.solr.languagemodels.model.DummyEmbeddingModel",
   "name": "dummy-1",
   "params": {
     "embedding": [1.0, 2.0, 3.0, 4.0],
