+++ solr/modules/language-models/src/test-files/textToVectorModelExamples/exception-throwing-model.json
@@ -1,0 +1,6 @@
+{
+  "class": "org.apache.solr.languagemodels.model.ExceptionThrowingEmbeddingModel",
+  "name": "exception-throwing-model",
+  "params": {
+  }
+}
