+++ solr/modules/language-models/src/test-files/textToVectorModelExamples/huggingface-model.json
