+++ solr/modules/language-models/src/test-files/textToVectorModelExamples/mistralai-model.json
