+++ solr/modules/language-models/src/test-files/textToVectorModelExamples/openai-model.json
