+++ solr/modules/language-models/src/test/org/apache/solr/languagemodels/TestLanguageModelBase.java
@@ -38,11 +38,12 @@ public class TestLanguageModelBase extends RestTestBase {
   protected static Path tmpSolrHome;
   protected static Path tmpConfDir;
 
-  public static final String EMBEDDING_MODEL_FILE_NAME = "_schema_text-to-vector-model-store.json";
+  public static final String TEXT_TO_VECTOR_MODEL_FILE_NAME =
+      "_schema_text-to-vector-model-store.json";
   protected static final String COLLECTION = "collection1";
   protected static final String CONF_DIR = COLLECTION + "/conf";
 
-  protected static Path embeddingModelStoreFile = null;
+  protected static Path textToVectorModelStoreFile = null;
 
   protected static String IDField = "id";
   protected static String vectorField = "vector";
@@ -61,17 +62,17 @@ public class TestLanguageModelBase extends RestTestBase {
     tmpSolrHome = createTempDir();
     tmpConfDir = tmpSolrHome.resolve(CONF_DIR);
     PathUtils.copyDirectory(TEST_PATH(), tmpSolrHome.toAbsolutePath());
-    final Path embeddingStore = tmpConfDir.resolve(EMBEDDING_MODEL_FILE_NAME);
+    final Path textToVectorStore = tmpConfDir.resolve(TEXT_TO_VECTOR_MODEL_FILE_NAME);
 
     if (isPersistent) {
-      embeddingModelStoreFile = embeddingStore;
+      textToVectorModelStoreFile = textToVectorStore;
     }
 
-    if (Files.exists(embeddingStore)) {
+    if (Files.exists(textToVectorStore)) {
       if (log.isInfoEnabled()) {
-        log.info("remove model store config file in {}", embeddingStore.toAbsolutePath());
+        log.info("remove model store config file in {}", textToVectorStore.toAbsolutePath());
       }
-      Files.delete(embeddingStore);
+      Files.delete(textToVectorStore);
     }
 
     System.setProperty("managed.schema.mutable", "true");
@@ -87,7 +88,8 @@ public class TestLanguageModelBase extends RestTestBase {
   }
 
   public static void loadModel(String fileName, String status) throws Exception {
-    final URL url = TestLanguageModelBase.class.getResource("/embeddingModelExamples/" + fileName);
+    final URL url =
+        TestLanguageModelBase.class.getResource("/textToVectorModelExamples/" + fileName);
     final String multipleModels = Files.readString(Path.of(url.toURI()), StandardCharsets.UTF_8);
 
     assertJPut(
@@ -97,7 +99,8 @@ public class TestLanguageModelBase extends RestTestBase {
   }
 
   public static void loadModel(String fileName) throws Exception {
-    final URL url = TestLanguageModelBase.class.getResource("/embeddingModelExamples/" + fileName);
+    final URL url =
+        TestLanguageModelBase.class.getResource("/textToVectorModelExamples/" + fileName);
     final String multipleModels = Files.readString(Path.of(url.toURI()), StandardCharsets.UTF_8);
 
     assertJPut(
