+++ solr/modules/language-models/src/test/org/apache/solr/languagemodels/model/DummyEmbeddingModelTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.languagemodels.textvectorisation.model;
+package org.apache.solr.languagemodels.model;
 
 import org.apache.solr.SolrTestCase;
 import org.junit.Test;
