+++ solr/modules/language-models/src/test/org/apache/solr/languagemodels/search/TextToVectorQParserTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.languagemodels.textvectorisation.search;
+package org.apache.solr.languagemodels.search;
 
 import java.util.Arrays;
 import java.util.Locale;
